refactor: move prefix_helpers to nexus.lib for tier compliance (#3951)
import-linter caught a five-tier architecture violation introduced by
the descendant-access refactor: nexus.services.namespace.descendant_access
imported nexus.bricks.rebac.cache._prefix_helpers, but services tier
must not depend on bricks. Move the helper to nexus.lib (a lower tier
that all callers can reach) and update imports across visibility.py,
enforcer.py, descendant_access.py, and the helper's tests.

Also tighten the type annotation on _resolve_paths_with_status() to
satisfy mypy's [no-any-return] check.

All 109 unit tests still pass.
